+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v7.10.7](https://github.com/lexicalunit/spellbot/releases/tag/v7.10.7) - 2021-12-21
+
 ### Changed
 
 - Adds errors to tracer spans rather than using alerting events.